+++ onmt/Factory.lua
@@ -1,9 +1,34 @@
 local Factory = torch.class('Factory')
 
 local options = {
-  {'-brnn', false, [[Use a bidirectional encoder.]]},
-  {'-dbrnn', false, [[Use a deep bidirectional encoder.]]},
-  {'-pdbrnn', false, [[Use pyramidal deep bidirectional encoder.]]}
+  {
+    '-brnn', false,
+    [[Use a bidirectional encoder.]],
+    {
+      structural = 0
+    }
+  },
+  {
+    '-dbrnn', false,
+    [[Use a deep bidirectional encoder.]],
+    {
+      structural = 0
+    }
+  },
+  {
+    '-pdbrnn', false,
+    [[Use a pyramidal deep bidirectional encoder.]],
+    {
+      structural = 0
+    }
+  },
+  {
+    '-attention', 'global',
+    [[Attention model.]],
+    {
+      enum = {'none', 'global'}
+    }
+  }
 }
 
 function Factory.declareOpts(cmd)
@@ -11,6 +36,7 @@ function Factory.declareOpts(cmd)
   onmt.BiEncoder.declareOpts(cmd)
   onmt.DBiEncoder.declareOpts(cmd)
   onmt.PDBiEncoder.declareOpts(cmd)
+  onmt.GlobalAttention.declareOpts(cmd)
 end
 
 -- Return effective embeddings size based on user options.
@@ -98,14 +124,6 @@ local function buildInputNetwork(opt, dicts, wordSizes, pretrainedWords, fixWord
   return inputNetwork
 end
 
-local function fixWordEmbeddings(model, fix)
-  model:apply(function(mod)
-    if torch.typename(mod) == 'onmt.WordEmbedding' then
-      mod:fixEmbeddings(fix)
-    end
-  end)
-end
-
 function Factory.getOutputSizes(dicts)
   local outputSizes = { dicts.words:size() }
   for i = 1, #dicts.features do
@@ -114,23 +132,31 @@ function Factory.getOutputSizes(dicts)
   return outputSizes
 end
 
-function Factory.buildEncoder(opt, inputNetwork)
+function Factory.buildEncoder(opt, inputNetwork, verbose)
 
   if opt.brnn then
-    _G.logger:info('   - Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
-                   opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    if verbose then
+      _G.logger:info('   - Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
+                     opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    end
     return onmt.BiEncoder.new(opt, inputNetwork)
   elseif opt.dbrnn then
-    _G.logger:info('   - Deep Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
-                   opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    if verbose then
+      _G.logger:info('   - Deep Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
+                     opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    end
     return onmt.DBiEncoder.new(opt, inputNetwork)
   elseif opt.pdbrnn then
-    _G.logger:info('   - Pyramidal Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
-                   opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    if verbose then
+      _G.logger:info('   - Pyramidal Bidirectional %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
+                     opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    end
     return onmt.PDBiEncoder.new(opt, inputNetwork)
   else
-    _G.logger:info('   - Simple %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
-                   opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    if verbose then
+      _G.logger:info('   - Simple %s Encoder: %d layers, rnn_size %d, dropout %0.1f',
+                     opt.rnn_type, opt.layers, opt.rnn_size, opt.dropout)
+    end
     return onmt.Encoder.new(opt, inputNetwork)
   end
 
@@ -140,21 +166,22 @@ function Factory.buildWordEncoder(opt, dicts, verbose)
   if verbose then
     _G.logger:info(' * Encoder:')
   end
+
   local inputNetwork
   if dicts then
     inputNetwork = buildInputNetwork(opt, dicts,
                                      opt.src_word_vec_size or opt.word_vec_size,
-                                     opt.pre_word_vecs_enc, opt.fix_word_vecs_enc,
+                                     opt.pre_word_vecs_enc, opt.fix_word_vecs_enc == 1,
                                      verbose)
   else
     inputNetwork = nn.Identity()
     inputNetwork.inputSize = opt.dimInputSize
   end
 
-  return Factory.buildEncoder(opt, inputNetwork)
+  return Factory.buildEncoder(opt, inputNetwork, verbose)
 end
 
-function Factory.loadEncoder(pretrained, clone, opt)
+function Factory.loadEncoder(pretrained, clone)
   if clone then
     pretrained = onmt.utils.Tensor.deepClone(pretrained)
   end
@@ -179,14 +206,10 @@ function Factory.loadEncoder(pretrained, clone, opt)
     end
   end
 
-  if opt then
-    fixWordEmbeddings(encoder, opt.fix_word_vecs_enc)
-  end
-
   return encoder
 end
 
-function Factory.buildDecoder(opt, inputNetwork, generator)
+function Factory.buildDecoder(opt, inputNetwork, generator, attnModel)
   local inputSize = inputNetwork.inputSize
 
   if opt.input_feed == 1 then
@@ -199,7 +222,7 @@ function Factory.buildDecoder(opt, inputNetwork, generator)
   end
   local rnn = RNN.new(opt.layers, inputSize, opt.rnn_size, opt.dropout, opt.residual, opt.dropout_input)
 
-  return onmt.Decoder.new(inputNetwork, rnn, generator, opt.input_feed == 1)
+  return onmt.Decoder.new(opt, inputNetwork, rnn, generator, opt.input_feed == 1, attnModel)
 end
 
 function Factory.buildWordDecoder(opt, dicts, verbose)
@@ -209,24 +232,21 @@ function Factory.buildWordDecoder(opt, dicts, verbose)
 
   local inputNetwork = buildInputNetwork(opt, dicts,
                                          opt.tgt_word_vec_size or opt.word_vec_size,
-                                         opt.pre_word_vecs_dec, opt.fix_word_vecs_dec,
+                                         opt.pre_word_vecs_dec, opt.fix_word_vecs_dec == 1,
                                          verbose)
 
   local generator = Factory.buildGenerator(opt.rnn_size, dicts)
+  local attnModel = Factory.buildAttention(opt)
 
-  return Factory.buildDecoder(opt, inputNetwork, generator)
+  return Factory.buildDecoder(opt, inputNetwork, generator, attnModel)
 end
 
-function Factory.loadDecoder(pretrained, clone, opt)
+function Factory.loadDecoder(pretrained, clone)
   if clone then
     pretrained = onmt.utils.Tensor.deepClone(pretrained)
   end
 
   local decoder = onmt.Decoder.load(pretrained)
-
-  if opt then
-    fixWordEmbeddings(decoder, opt.fix_word_vecs_dec)
-  end
 
   return decoder
 end
@@ -236,6 +256,16 @@ function Factory.buildGenerator(rnnSize, dicts)
     return onmt.FeaturesGenerator(rnnSize, Factory.getOutputSizes(dicts))
   else
     return onmt.Generator(rnnSize, dicts.words:size())
+  end
+end
+
+function Factory.buildAttention(args)
+  if args.attention == 'none' then
+    _G.logger:info('   - No Attention')
+    return onmt.NoAttention(args, args.rnn_size)
+  else
+    _G.logger:info('   - Global Attention: '..args.global_attention)
+    return onmt.GlobalAttention(args, args.rnn_size)
   end
 end