+++ onmt/train/Checkpoint.lua
@@ -2,9 +2,17 @@
 local Checkpoint = torch.class('Checkpoint')
 
 local options = {
-  {'-train_from', '',  [[If training from a checkpoint then this is the path to the pretrained model.]],
-                         {valid=onmt.utils.ExtendedCmdLine.fileNullOrExists}},
-  {'-continue', false, [[If training from a checkpoint, whether to continue the training in the same configuration or not.]]}
+  {
+    '-train_from', '',
+    [[Path to a checkpoint.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.fileNullOrExists
+    }
+  },
+  {
+    '-continue', false,
+    [[If set, continue the training where it left off.]]
+  }
 }
 
 function Checkpoint.declareOpts(cmd)
@@ -23,6 +31,7 @@ end
 function Checkpoint:save(filePath, info)
   info.learningRate = self.optim:getLearningRate()
   info.optimStates = self.optim:getStates()
+  info.rngStates = onmt.utils.Cuda.getRNGStates()
 
   local data = {
     models = {},
@@ -78,31 +87,62 @@ end
 
 function Checkpoint.loadFromCheckpoint(opt)
   local checkpoint = {}
+  local paramChanges = {}
+
   if opt.train_from:len() > 0 then
     _G.logger:info('Loading checkpoint \'' .. opt.train_from .. '\'...')
 
     checkpoint = torch.load(opt.train_from)
 
-    opt.layers = checkpoint.options.layers
-    opt.rnn_size = checkpoint.options.rnn_size
-    opt.brnn = checkpoint.options.brnn
-    opt.brnn_merge = checkpoint.options.brnn_merge
-    opt.input_feed = checkpoint.options.input_feed
+    local function restoreOption(name)
+      if checkpoint.options[name] ~= nil then
+        opt[name] = checkpoint.options[name]
+      end
+    end
 
-    -- Resume training from checkpoint
+    if checkpoint.info.rngStates and not (opt._is_default['seed'] or opt.seed == checkpoint.options.seed) then
+      _G.logger:warning('\'-seed\' value has changed - ignoring saved rng states')
+      checkpoint.info.rngStates = nil
+    end
+
+    if opt.continue and (not (opt._is_default['learning_rate'] or opt.learning_rate == checkpoint.options.learning_rate) or
+        not (opt._is_default['start_epoch'] or opt.start_epoch == checkpoint.options.start_epoch)) then
+      _G.logger:warning('\'-continue\' option is used but learning_rate or start_epoch are set and different than previous epoch. Ignoring \'-continue\'')
+      opt.continue = nil
+    end
+
+    -- Reload and check options.
+    for k, v in pairs(opt) do
+      if k:sub(1, 1) ~= '_' then
+
+        if opt.continue and opt._train_state[k] then
+          -- Training states should be retrieved when continuing a training.
+          restoreOption(k)
+        elseif opt._structural[k] or opt._init_only[k] then
+          -- If an option was set by the user, check that we can actually change it.
+          local valueChanged = not opt._is_default[k] and v ~= checkpoint.options[k]
+
+          if valueChanged then
+            if opt._init_only[k] then
+              _G.logger:warning('Cannot change initialization option -%s. Ignoring.', k)
+              restoreOption(k)
+            elseif opt._structural[k] and opt._structural[k] == 0 then
+              _G.logger:warning('Cannot change dynamically option -%s. Ignoring.', k)
+              restoreOption(k)
+            elseif opt._structural[k] and opt._structural[k] == 1 then
+              paramChanges[k] = v
+            end
+          else
+            restoreOption(k)
+          end
+
+        end
+
+      end
+    end
+
     if opt.continue then
-      opt.fix_word_vecs_enc = checkpoint.options.fix_word_vecs_enc
-      opt.fix_word_vecs_dec = checkpoint.options.fix_word_vecs_dec
-
-      opt.optim = checkpoint.options.optim
-      opt.learning_rate_decay = checkpoint.options.learning_rate_decay
-      opt.start_decay_at = checkpoint.options.start_decay_at
-      opt.curriculum = checkpoint.options.curriculum
-
-      opt.decay = checkpoint.options.decay or opt.decay
-      opt.min_learning_rate = checkpoint.options.min_learning_rate or opt.min_learning_rate
-      opt.start_decay_ppl_delta = checkpoint.options.start_decay_ppl_delta or opt.start_decay_ppl_delta
-
+      -- When continuing, some options are initialized with their last known value.
       opt.learning_rate = checkpoint.info.learningRate
       opt.start_epoch = checkpoint.info.epoch
       opt.start_iteration = checkpoint.info.iteration
@@ -110,10 +150,14 @@ function Checkpoint.loadFromCheckpoint(opt)
       _G.logger:info('Resuming training from epoch ' .. opt.start_epoch
                          .. ' at iteration ' .. opt.start_iteration .. '...')
     else
-      checkpoint.info = nil
+      -- Otherwise, we can drop previous training information.
+      checkpoint.info.learningRate = nil
+      checkpoint.info.epoch = nil
+      checkpoint.info.iteration = nil
     end
   end
-  return checkpoint, opt
+
+  return checkpoint, opt, paramChanges
 end
 
 return Checkpoint