+++ onmt/train/Optim.lua
@@ -54,25 +54,82 @@ end
 local Optim = torch.class('Optim')
 
 local options = {
-  {'-max_batch_size',     64   , [[Maximum batch size]],
-                                 {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-uneven_batches',     false, [[If true, batches are filled up to max_batch_size even if they have different sizes.]]},
-  {'-optim',              'sgd', [[Optimization method.]],
-                                 {enum={'sgd', 'adagrad', 'adadelta', 'adam'}}},
-  {'-learning_rate',       1   , [[Starting learning rate. If adagrad or adam is used,
-                                      then this is the global learning rate. Recommended settings are: sgd = 1,
-                                      adagrad = 0.1, adam = 0.0002]]},
-  {'-min_learning_rate',   0   , [[Do not continue the training past this learning rate]]},
-  {'-max_grad_norm',       5   , [[If the norm of the gradient vector exceeds this renormalize it to have
-                                       the norm equal to max_grad_norm]]},
-  {'-learning_rate_decay', 0.5 , [[Learning rate decay factor]]},
-  {'-start_decay_at',      9   , [[With 'default' decay mode, start decay after this epoch]],
-                                 {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-start_decay_ppl_delta', 0 , [[Start decay when validation perplexity improvement is lower than this value]]},
-  {'-decay',          'default', [[When to apply learning rate decay.
-                                 'default': decay after each epoch past start_decay_at or as soon as the validation perplexity is not improving more than start_decay_ppl_delta,
-                                 'perplexity_only': only decay when validation perplexity is not improving more than start_decay_ppl_delta]],
-                                 {enum={'default', 'perplexity_only'}}}
+  {
+    '-max_batch_size', 64,
+    [[Maximum batch size.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt()
+    }
+  },
+  {
+    '-uneven_batches', false,
+    [[If set, batches are filled up to max_batch_size even if source lengths are different.
+      Slower but needed for some tasks.]]
+  },
+  {
+    '-optim', 'sgd',
+    [[Optimization method.]],
+    {
+      enum = {'sgd', 'adagrad', 'adadelta', 'adam'},
+      train_state = true
+    }
+  },
+  {
+    '-learning_rate', 1,
+    [[Starting learning rate. If adagrad or adam is used, then this is the global learning rate.
+      Recommended settings are: sgd = 1, adagrad = 0.1, adam = 0.0002.]],
+    {
+      train_state = true
+    }
+  },
+  {
+    '-min_learning_rate', 0,
+    [[Do not continue the training past this learning rate value.]],
+    {
+      train_state = true
+    }
+  },
+  {
+    '-max_grad_norm', 5,
+    [[Clip the gradients norm to this value.]],
+    {
+      train_state = true
+    }
+  },
+  {
+    '-learning_rate_decay', 0.7,
+    [[Learning rate decay factor: `learning_rate = learning_rate * learning_rate_decay`.]],
+    {
+      train_state = true
+    }
+  },
+  {
+    '-start_decay_at', 9,
+    [[In "default" decay mode, start decay after this epoch.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt(),
+      train_state = true
+    }
+  },
+  {
+    '-start_decay_ppl_delta', 0,
+    [[Start decay when validation perplexity improvement is lower than this value.]],
+    {
+      train_state = true
+    }
+  },
+  {
+    '-decay', 'default',
+    [[When to apply learning rate decay.
+      `default`: decay after each epoch past `-start_decay_at` or as soon as the
+      validation perplexity is not improving more than `-start_decay_ppl_delta`,
+      `perplexity_only`: only decay when validation perplexity is not improving more than
+      `-start_decay_ppl_delta`.]],
+    {
+      enum = {'default', 'perplexity_only'},
+      train_state = true
+    }
+  }
 }
 
 function Optim.declareOpts(cmd)