+++ onmt/translate/DecoderAdvancer.lua
@@ -17,12 +17,15 @@ Parameters:
   * `updateSeqLengthFunc` - optional, sequence length adaptation function after encoder
 
 --]]
-function DecoderAdvancer:__init(decoder, batch, context, max_sent_length, max_num_unks, decStates, dicts, updateSeqLengthFunc)
+function DecoderAdvancer:__init(decoder, batch, context, max_sent_length, max_num_unks, decStates, dicts, length_norm, coverage_norm, eos_norm, updateSeqLengthFunc)
   self.decoder = decoder
   self.batch = batch
   self.context = context
   self.max_sent_length = max_sent_length or math.huge
   self.max_num_unks = max_num_unks or math.huge
+  self.length_norm = length_norm or 0.0
+  self.coverage_norm = coverage_norm or 0.0
+  self.eos_norm = eos_norm or 0.0
   self.decStates = decStates or onmt.utils.Tensor.initTensorTable(
     decoder.args.numEffectiveLayers,
     onmt.utils.Cuda.convert(torch.Tensor()),
@@ -47,11 +50,23 @@ function DecoderAdvancer:initBeam()
     end
   end
   local sourceSizes = onmt.utils.Cuda.convert(self.batch.sourceSize)
+  local attnProba = torch.FloatTensor(self.batch.size, self.context:size(2)):fill(0.0001)
+  -- Mask padding
+  for i = 1,self.batch.size do
+    local pad_size = self.context:size(2) - sourceSizes[i]
+    if (pad_size ~= 0) then
+      attnProba[{ i, {1,pad_size} }] = 1.0
+    end
+  end
 
   -- Define state to be { decoder states, decoder output, context,
-  -- attentions, features, sourceSizes, step }.
-  local state = { self.decStates, nil, self.context, nil, features, sourceSizes, 1 }
-  return onmt.translate.Beam.new(tokens, state)
+  -- attentions, features, sourceSizes, step, cumulated attention probablities }.
+  local state = { self.decStates, nil, self.context, nil, features, sourceSizes, 1, attnProba }
+  local params = {}
+  params.length_norm = self.length_norm
+  params.coverage_norm = self.coverage_norm
+  params.eos_norm = self.eos_norm
+  return onmt.translate.Beam.new(tokens, state, params)
 end
 
 --[[Updates beam states given new tokens.
@@ -63,8 +78,8 @@ Parameters:
 ]]
 function DecoderAdvancer:update(beam)
   local state = beam:getState()
-  local decStates, decOut, context, _, features, sourceSizes, t
-    = table.unpack(state, 1, 7)
+  local decStates, decOut, context, _, features, sourceSizes, t, cumAttnProba
+    = table.unpack(state, 1, 8)
   local tokens = beam:getTokens()
   local token = tokens[#tokens]
   local inputs
@@ -86,7 +101,10 @@ function DecoderAdvancer:update(beam)
   decOut, decStates = self.decoder:forwardOne(inputs, decStates, context, decOut)
   t = t + 1
   local softmaxOut = self.decoder.softmaxAttn.output
-  local nextState = {decStates, decOut, context, softmaxOut, nil, sourceSizes, t}
+
+  cumAttnProba = cumAttnProba:typeAs(softmaxOut):add(softmaxOut)
+
+  local nextState = {decStates, decOut, context, softmaxOut, nil, sourceSizes, t, cumAttnProba}
   beam:setState(nextState)
 end